+++ HISTORY.rst
@@ -2,7 +2,15 @@
 History
 =======
 
-0.1.0 (2020-03-17)
+0.1.1 (2020-11-9)
+-----------------
+
+* Mixed lubrication api brought into line with other steps
+* bug fixes in mixed lubrication solver
+* added just in time compilation to unified reynolds solver
+* converted IterSemiSystem to quasi static
+
+0.1.0 (2020-11-7)
 ------------------
 
 * First release on PyPI.